Fix some coding mistakes spotted by static code analyzer

+++ source/game/g_ascript.cpp
@@ -1780,7 +1780,7 @@ static unsigned int objectGameClient_getPMoveFeatures( gclient_t *self )
 static void objectGameClient_setPMoveMaxSpeed( float speed, gclient_t *self )
 {
 	if( speed < 0.0f )
-		self->ps.pmove.stats[PM_STAT_MAXSPEED] = DEFAULT_PLAYERSPEED;
+		self->ps.pmove.stats[PM_STAT_MAXSPEED] = (short)DEFAULT_PLAYERSPEED;
 	else
 		self->ps.pmove.stats[PM_STAT_MAXSPEED] = ( (int)speed & 0xFFFF );
 }
@@ -1793,7 +1793,7 @@ static float objectGameClient_getPMoveMaxSpeed( gclient_t *self )
 static void objectGameClient_setPMoveJumpSpeed( float speed, gclient_t *self )
 {
 	if( speed < 0.0f )
-		self->ps.pmove.stats[PM_STAT_JUMPSPEED] = DEFAULT_JUMPSPEED;
+		self->ps.pmove.stats[PM_STAT_JUMPSPEED] = (short)DEFAULT_JUMPSPEED;
 	else
 		self->ps.pmove.stats[PM_STAT_JUMPSPEED] = ( (int)speed & 0xFFFF );
 }
@@ -1806,7 +1806,7 @@ static float objectGameClient_getPMoveJumpSpeed( gclient_t *self )
 static void objectGameClient_setPMoveDashSpeed( float speed, gclient_t *self )
 {
 	if( speed < 0.0f )
-		self->ps.pmove.stats[PM_STAT_DASHSPEED] = DEFAULT_DASHSPEED;
+		self->ps.pmove.stats[PM_STAT_DASHSPEED] = (short)DEFAULT_DASHSPEED;
 	else
 		self->ps.pmove.stats[PM_STAT_DASHSPEED] = ( (int)speed & 0xFFFF );
 }
